requirements.txt
allure-pytest
jsonschema
requests
pytest
jsonref